Fix vsphere region enumeration
This commit fixes the issue with vSphere region enumeration data
source and also updates the SDK to 0.2.8.

The fix is to set the ID of the data source to hostname instead
of dcid which is optional for vRA on-prem.

Latest SDK v0.2.8 has change to security group model that combines
both ingress and egress into a single set of rules.

Issues fixed: #139

Testing done: Manual with vRA 8.0.1 HF1 and HF2, 8.1 and vRA Cloud

Signed-off-by: Deepak Mettem <dmettem@vmware.com>