Fix error with deployment requests from catalog item
With the introduction of bulk request feature in 8.1, the response model for
the catalog API to request a deployment from a catalog item is updated.  This
change introduced a new API version '2020-01-30'.

This change is to request deployment from catalog item on the apiVersion
'2019-01-15' to make sure all vRA 8.x versions are supported with the provider.

Signed-off-by: Deepak Mettem <dmettem@vmware.com>