Persist state for deployments that failed mid-flight
For deployments failed mid-flight, state is not persisted correctly causing
the refresh to fail. This commit ensures that a deployment id is persisted
during create, so that even if the deployment fails or timesout, the
deployment resource is still created in the state and refresh works.

In case the provider times out before the deployment request completes i.e
successful or failed, refresh will pull the latest state of deployment.
For in-place updates of such deployment resources that timed out during
create should be untainted first.

Signed-off-by: Deepak Mettem <dmettem@vmware.com>