Show failure message if deployment fails and add update statuses as valid
Currently, if a deployment fails during create, provider reports that
deployment resource is added, however the state file has the correct
status i.e. CREATE_FAILED.

With this change, if a deployment is failed during create, the apply
command shows the failure message and also the state file is updated
with current state of deployment.

Also added UPDATE_INPROGRESS, UPDATE_SUCCESSFUL and UPDATE_FAILEED as
valid statues in deploymentCreateStatusRefreshFunc.

Signed-off-by: Deepak Mettem <dmettem@vmware.com>