Fix incorrect deployment update plans
If a deployment is requested only with required inputs, upon successful
creation of deployment, the state file includes inputs not only that are
provided by the user in the config file, but also those with default
values in the blueprint/catalog item. Hence, when the plan/apply command
is run next time, the plan shows that the deployment state is changed and
it needs to update those deployment inputs that has default value with null.

This change introduces a new attribute 'inputs_including_defaults' with all
the inputs, i.e. both user provided and those with default values but not
provided by the user. The existing attribute 'inputs' includes only the
user provided inputs. These inputs provided by the user in the config file
are updated when deployment is updated in vRA directly and the terraform
state file is out of sync.

Signed-off-by: Deepak Mettem <dmettem@vmware.com>