Fix #219 by allowing deployment resource updates (#220)
Before this change, the only updates allowed on  resource are changes
to metadata (i.e. name and description) and inputs (i.e. update day-2
action). When there are any changes to blueprint_id, blueprint_version,
blueprint_content, catalog_item_id, or catalog_item_version, TF shows
the in-place update as part of the plan, however, the changes are not
made to the resource, thus by showing the same plan over and over.

This change introduces the Cloud Assembly Update existing deployment
feature where the user wants to update an existing deployment with a
different blueprint i.e. either a different blueprint_id or
blueprint_version or even blueprint_content (inline blueprint).
This is in-place update of the deployment where the deployment id
doesn't change but only the resources in the deployment are changed.

As the Service Broker doesn't let the user update an existing deployment
with a different catalog_item_id or catalog_item_version, this change
forces a new deployment resource thus by reports the plan correctly when
either of these are changed in the configuraion. This is destroy and
re-create of the deployment.

This change also updates the terraform-plugin-sdk to the latest version
1.15.0.

Signed-off-by: Deepak Mettem <dmettem@vmware.com>