Add support for deployment request inputs of array and object types
This commit adds the support for  of array and object types in
`vra_deployment` resource.

By default, all `TypeMap` properties in resource schema supports only
primitive types and the values should be of type `string`  even when the
`Elem` property is not specified. This is the same behavior even with
terraform-plugin-sdk v2.

Hence, the inputs of 'array' and 'object' types from vRA cloud templates
in `vra_deployment` resource should be provided in string format in
Terraform configuration files.

Here's an example on how the inputs of different types can be specified
in configuration file:

inputs = {
    StringInput  = small
    integerInput = 1
    booleanInput = false
    numberInput  = 10.0
    arrayInput   = jsonencode([foo, bar, where, waldo])
    objectInput  = jsonencode({ key1 : value1, key2 : [1, 2, 3, 4] })
}

The state includes `inputs` and `inputs_including_defaults`, out of which
`inputs` are only those that are included in configuration file and
`inputs_including_defaults` are both those from the configuration file and
any inputs omitted from configuration file but have defaults defined in
cloud templates.

Fixes #205.

Signed-off-by: Deepak Mettem <dmettem@vmware.com>