Allow vra_machine creation with first class vSphere disks
Due to limitations to attach a first class vSphere disk to a machine during
provisioning as day-0 operation, the vra_machine resource create operation is
now changed to create the machine first without disks (even if given in the
configuration file) and then the disks are attached one by one as day-2 actions.

Also, the file name for vra_fabric_network_vsphere resource is modified to be
in consistent with other file names.

Signed-off-by: Deepak Mettem <dmettem@vmware.com>