Create getRefreshToken function
This commit creates the GetRefreshToken that is used by the provider module to retrieve and set a refresh token if one is not provided by the user, but a username and password is.

As noted by the comments in the file, the Swagger spec does not properly populate the `/csp/gateway/am/api/login` endpoint, and so we can't use the `vra-go-sdk` to generate an `apiclient` like the other functions, so I'm falling back to relying on the default `net/http` to execute the API call and fetch the refresh token.

Signed-off-by: Marcus Puckett <mpuckett259@gmail.com>