Update provider to handle user/pass/domain
This commit updates the provider module to read in the username, password, and domain properties as configured by the provider in the TF file, and then uses the new GetRefreshToken function to fetch a refresh token if a username and password is provided, and no tokens are provided. Since the domain isn't required, that isn't checked for and just blindly passed in to be validated in the function.

Signed-off-by: Marcus Puckett <mpuckett259@gmail.com>