Simplify refresh token function
This commit simplifies the refresh token fetch call. The domain body can be an empty string or specify a domain if required. If the domain is incorrect it will return a 400 Bad Request error.

This commit also sets the timeout to 5 seconds as per general good Go practices.

I'm also changing to defining a request and passing it to the httpClient.Do() function now so I can add headers as needed. Probably not required but I wanted to set the Contet-Type and Accept headers and the httpClient.Post() only lets you set Content-Type.