Return all deployment resources
The API to get the deployment resources returns a pageable response.
By default, each page contains 20 resources.

Unfortunatelly, the client generated from the swagger specification
does not allow to retrieve a specific page, so we can only retrieve
the first page, leading to an incomplete list of resources when the
deployment contains more than 20 resources.

As a workaround until the client supports proper pagination, we will
increase the number of resources to be returned per page to 1000.

Signed-off-by: Ferran Rodenas <rodenasf@vmware.com>