fix: use `expandInputsToString` for deployment inputs
Replaces `expandInputs` with `expandInputsToString` when setting the 'inputs' field in `dataSourceDeploymentRead`. This likely ensures the deployment inputs are processed as strings.

Otherwise, using `expandInputs()` instead of `expandInputsToString()` when setting deployment inputs, which causes type conversion errors when Automation returns non-string values.

It's worth noting this this was addressed in`r/vra_deployment` in PR #259 to use `expandInputsToString()` but it was never applied to the data source.

Refs:
- #605
- #493

Signed-off-by: Ryan Johnson <ryan.johnson@broadcom.com>